Move callable to its own header
