+++ src/autowiring/CMakeLists.txt
@@ -64,6 +64,7 @@ set(Autowiring_SRCS
   Bolt.h
   BoltBase.cpp
   BoltBase.h
+  callable.h
   CallExtractor.h
   CallExtractor.cpp
   ContextEnumerator.cpp
@@ -79,6 +80,7 @@ set(Autowiring_SRCS
   CoreJob.h
   CoreObject.cpp
   CoreObject.h
+  CoreObjectDescriptor.h
   CoreRunnable.cpp
   CoreRunnable.h
   CoreThread.cpp
@@ -141,7 +143,6 @@ set(Autowiring_SRCS
   once.cpp
   Parallel.h
   Parallel.cpp
-  CoreObjectDescriptor.h
   SatCounter.h
   SatCounter.cpp
   SlotInformation.cpp
